Check whether logs dir exists before cleaning

+++ app/src/main/java/io/xpipe/app/core/AppLogs.java
@@ -98,7 +98,7 @@ public class AppLogs {
     public static void init() {
         var logDir = AppProperties.get().getDataDir().resolve("logs");
 
-        if (XPipeSession.get().isNewBuildSession()) {
+        if (XPipeSession.get().isNewBuildSession() && Files.exists(logDir)) {
             try {
                 FileUtils.cleanDirectory(logDir.toFile());
             } catch (Exception ex) {
